update home
 component divided / responsivness
