+++ src/data.json
@@ -1,0 +1,10 @@
+[{"response":"C","no":"084/C/NBP/2023","tradingDate":"2023-04-28","effectiveDate":"2023-05-02",
+"data":[{"date":"16.04.2023","type":"expense","category":"car","comment":"comment","sum":177.52,"":"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52 , "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"},
+{"date":"16.04.2023","type":"expense","category":"car","comment":"comment", "sum":177.52, "" :"edit butt / X"}
+]}]
