--- src/pages/Home.module.scss
@@ -1,1 +1,0 @@
-@use "../stylesheets/vars" as *;
