+++ src/pages/IconSvg.jsx
@@ -2,8 +2,8 @@ import sprite from "../images/vectors/icons.svg"
 
 const IconSvg = ({icon}) =>{
     return<>
-    <svg>
-        <use href={sprite + '#icon-' +icon } />
+    <svg className={icon}>
+        <use className={icon+"_Use"}  href={sprite + '#icon-' +icon } />
     </svg>
     </>
 }
