+++ src/pages/home/AddButton.jsx
@@ -1,0 +1,13 @@
+import scss from './Home.module.scss';
+
+const AddButton =() =>{
+    return (
+        <div className={scss.statsButtonContainer}>        
+            <button className={scss.statsButton}>
+        +
+        </button>
+        </div>
+    )
+}
+
+export default AddButton;
