+++ src/pages/home/Home.module.scss
@@ -1,0 +1,1 @@
+@use "../../stylesheets/vars" as *;
