+++ src/stylesheets/_vars.scss
@@ -7,4 +7,4 @@
 
 // components
 @use './components/global' as *;
-@use '../pages/home' as *;
+@use './components/home' as *;
